requirements.txt
# Dev environment
pip
autopep8

# App
flask
flask-sqlalchemy
sqlalchemy
python-telegram-bot
APScheduler
pytz
pyTelegramBotAPI
schedule
mcstatus
gunicorn
python-dotenv
